Moved EmptySupportRecyclerView.java
