+++ app/src/main/res/layout/fragment_advertise_coupon.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     tools:context="jp.ac.it_college.std.ikemen.reachable.company.coupon.AdvertiseCouponFragment">
 
-    <jp.ac.it_college.std.ikemen.reachable.company.EmptySupportRecyclerView
+    <jp.ac.it_college.std.ikemen.reachable.company.view.EmptySupportRecyclerView
         android:id="@+id/coupon_list"
         android:scrollbars="vertical"
         android:layout_width="match_parent"
