+++ app/src/main/res/layout/fragment_coupon_select.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     tools:context=".coupon.CouponSelectFragment">
 
-    <jp.ac.it_college.std.ikemen.reachable.company.EmptySupportRecyclerView
+    <jp.ac.it_college.std.ikemen.reachable.company.view.EmptySupportRecyclerView
         android:id="@+id/coupon_list"
         android:scrollbars="vertical"
         android:layout_width="match_parent"
